+++ CHANGELOG.md
@@ -15,6 +15,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   of characters (disabled by default).
 - Add support for `--literal` from [PanGan21](https://github.com/PanGan21)
 
+### Fixed
+
+### Changed
+- Human readable size flag `-h` re-enabled. [sdorr0](https://github.com/sdorr0)
+
 ## [v1.0.0] - 2023-08-25
 
 ### Added